Some cleanup, renaming
