+++ files/usr/lib/cinnamon-settings/modules/cs_panel.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python
 
-import settings
-from settings import *
+from SettingsWidgets import *
 
 class Module:
     def __init__(self, content_box):
